RPC: Move get_transaction_by_hash function from mempool into blockchain dispatcher
Seed nodes do not have a mempool, because they do not run a validator,
so this change makes the method accessible on the devnet seed node.
