+++ rpc-interface/src/mempool.rs
@@ -1,6 +1,6 @@
 use async_trait::async_trait;
 
-use crate::types::{HashOrTx, MempoolInfo, Transaction};
+use crate::types::{HashOrTx, MempoolInfo};
 use nimiq_hash::Blake2bHash;
 
 #[nimiq_jsonrpc_derive::proxy(name = "MempoolProxy", rename_all = "camelCase")]
@@ -9,12 +9,6 @@ pub trait MempoolInterface {
     type Error;
 
     async fn push_transaction(&mut self, raw_tx: String) -> Result<Blake2bHash, Self::Error>;
-
-    async fn get_transaction_by_hash(
-        &mut self,
-        hash: Blake2bHash,
-        check_mempool: Option<bool>,
-    ) -> Result<Transaction, Self::Error>;
 
     async fn mempool_content(
         &mut self,
