+++ rpc-server/src/dispatchers/mempool.rs
@@ -3,12 +3,11 @@ use std::sync::Arc;
 use async_trait::async_trait;
 use beserial::Deserialize;
 
-use nimiq_blockchain::AbstractBlockchain;
 use nimiq_hash::{Blake2bHash, Hash};
 use nimiq_mempool::mempool::Mempool;
 
 use nimiq_rpc_interface::mempool::MempoolInterface;
-use nimiq_rpc_interface::types::{HashOrTx, MempoolInfo, Transaction};
+use nimiq_rpc_interface::types::{HashOrTx, MempoolInfo};
 
 use crate::error::Error;
 
@@ -38,53 +37,6 @@ impl MempoolInterface for MempoolDispatcher {
             Ok(_) => Ok(txid),
             Err(e) => Err(Error::MempoolError(e)),
         }
-    }
-
-    /// Tries to fetch a transaction (including reward transactions) given its hash. It has an option
-    /// to also search the mempool for the transaction, it defaults to false.
-    async fn get_transaction_by_hash(
-        &mut self,
-        hash: Blake2bHash,
-        check_mempool: Option<bool>,
-    ) -> Result<Transaction, Error> {
-        // First check the mempool.
-        if check_mempool == Some(true) {
-            if let Some(tx) = self.mempool.get_transaction_by_hash(&hash) {
-                return Ok(Transaction::from_transaction(tx));
-            }
-        }
-
-        // Now check the blockchain.
-        let blockchain = self.mempool.blockchain.read();
-
-        // Get all the extended transactions that correspond to this hash.
-        let mut extended_tx_vec = blockchain.history_store.get_ext_tx_by_hash(&hash, None);
-
-        // Unpack the transaction or raise an error.
-        let extended_tx = match extended_tx_vec.len() {
-            0 => {
-                return Err(Error::TransactionNotFound(hash));
-            }
-            1 => extended_tx_vec.pop().unwrap(),
-            _ => {
-                return Err(Error::MultipleTransactionsFound(hash));
-            }
-        };
-
-        // Convert the extended transaction into a regular transaction. This will also convert
-        // reward inherents.
-        let block_number = extended_tx.block_number;
-        let timestamp = extended_tx.block_time;
-
-        return match extended_tx.into_transaction() {
-            Ok(tx) => Ok(Transaction::from_blockchain(
-                tx,
-                block_number,
-                timestamp,
-                blockchain.block_number(),
-            )),
-            Err(_) => Err(Error::TransactionNotFound(hash)),
-        };
     }
 
     async fn mempool_content(
